emplRepository moved to domain catalog
